+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Added
 
 - Enable termination events for CAPZ node pools.
+- Add support for creating workload cluster keypairs using the `login` command.
 
 ## [1.41.1] - 2021-10-04